Warnings are now considered errors at build time, printf now uses the format attribute for compile time format specifier checking, fixed bug in printf when printing integers larger then 32 bits

+++ kernel/src/string/printf.c
@@ -17,7 +17,7 @@ int printf(const char *restrict format, ...)
 
 	va_end(args);
 
-	for (int i = 0; i < written; i++)
+	for (size_t i = 0; i < written; i++)
 	{
 		serial_write(buffer[i]);
 	}
